fix(sidecar): keep stub host rpc non-routable
