+++ llms.txt
@@ -694,7 +694,7 @@ tools:
 - Pass `probe_tool="maya_diagnostics__ping"` (or CLI `--probe-tool`) to require one real `tools/call` through the sidecar before reporting ready; optional `probe_arguments` / `--probe-args-json` must be a JSON object. Probe failures keep polling until timeout in `wait_for_sidecar_ready(...)`.
 - `stop_runtime_entries(..., role="per-dcc-sidecar", include_host_processes=False)` requests sidecar stop without terminating the owning DCC host by default.
 - `build_sidecar_command(...)` returns a canonical `dcc-mcp-server sidecar` argv + environment updates for DCC startup hooks.
-- `launch_sidecar(...)` starts that sidecar with detached stdio by default so the DCC main process is not blocked; treat the registry row as dispatch-ready only after `metadata.dispatch_status=ready` and `metadata.mcp_url` appear. Host RPC failures stay visible with `metadata.dispatch_status=unavailable` plus `failure_stage` / `failure_reason`; they may still publish `metadata.mcp_url` for diagnostics, but `/v1/readyz` reports `dispatcher=false` and `tools/call` returns a structured transport error until dispatch becomes ready.
+- `launch_sidecar(...)` starts that sidecar with detached stdio by default so the DCC main process is not blocked; treat the registry row as dispatch-ready only after `metadata.dispatch_status=ready` and `metadata.mcp_url` appear. Host RPC failures stay visible with `metadata.dispatch_status=unavailable` plus `failure_stage` / `failure_reason`; they may still publish `metadata.mcp_url` for diagnostics, but `/v1/readyz` reports `dispatcher=false` and `tools/call` returns a structured transport error until dispatch becomes ready. `stub://` is test-only and remains unavailable by default even when it connects, so adapter startup checks must use a real host RPC scheme.
 - `inspect_install_root(path)` reports loaded native artifacts under the install root as `status="requires_restart"` with `locked_path`.
 - `safe_remove_tree(path)` / `safe_replace_tree(source, destination)` attempt immediate cleanup and classify Windows native-file lock failures into a deferred operation.
 - CLI: `python -m dcc_mcp_core.install_lifecycle {query|stop|layout|sidecar-command|launch-sidecar|sidecar-ready|plan-update|inspect|remove|replace}`.
