+++ skills/dcc-mcp-creator/SKILL.md
@@ -51,7 +51,7 @@ skill taxonomy), load `dcc-mcp-skills-creator` instead.
 5. Keep DCC identity data-driven: `dcc_name`, `server_name`, env-var prefix, skill names, and gateway metadata.
 6. Use core helpers for skill discovery, `MinimalModeConfig`, project tools, resources, diagnostics, context snapshots, install lifecycle, and gateway failover before writing adapter-local wrappers.
 7. Choose the `dcc-mcp-server` run mode deliberately: no subcommand, `auto`, and `serve` ensure the machine-wide gateway daemon, register this process as a backend, and keep a lightweight guardian running after registration; `serve --no-auto-gateway` is per-DCC only, `auto --legacy-gateway-election` restores the old embedded first-wins gateway, and `gateway` runs the machine-wide daemon with no inline DCC execution. Python `DccServerBase` adapters, Rust `dcc-mcp-server sidecar` processes with `gateway_port > 0`, and registered `dcc-mcp-server translate` bridges follow the same daemon-first contract at startup and keep the same daemon guardian running in daemon-backed mode so a surviving DCC or bridge can re-ensure the daemon after `/health` disappears; the shared `gateway-launch.lock` is reclaimed after `DCC_MCP_GATEWAY_LAUNCH_LOCK_STALE_SECS` if a launcher crashes mid-startup, so adapters should share `DCC_MCP_REGISTRY_DIR` and not delete the lock manually. `dcc_diagnostics__gateway_failover` should report `daemon-backed` or `embedded-fallback` so operators can distinguish the mode.
-8. When a DCC startup hook must spawn the per-DCC sidecar without importing native core, use `dcc_mcp_core.install_lifecycle.build_sidecar_command(...)` or `launch_sidecar(...)`; they produce the canonical `dcc-mcp-server sidecar` argv, keep stdio detached by default, stamp the shared registry/gateway env, and let the child ensure the daemon without blocking the DCC main process unless `no_ensure_gateway=True` or `legacy_gateway_election=True`. Do not treat a `per-dcc-sidecar` registry row alone as proof that tools are callable: the adapter must expose a supported host RPC bridge to its DCC dispatcher/skills, and startup code or installers should use `sidecar_readiness_status(...)` / `wait_for_sidecar_ready(...)` for a bounded import-light verdict before claiming the plugin is ready. Pass `probe_tool="<dcc>_diagnostics__ping"` (or CLI `--probe-tool`) when the adapter needs proof that one real `tools/call` reaches the host dispatcher; probe failures keep the verdict unready until timeout. Operators should see `dispatch_ready=True` from those helpers, `query_runtime_state(...)`, `GET /admin/api/workers`, or raw `metadata.dispatch_status=ready` plus `metadata.mcp_url`; failed host RPC startup stays registered with `dispatch_ready=False`, `metadata.dispatch_status=unavailable`, and `failure_stage` / `failure_reason` for diagnostics. A failed sidecar may still publish `metadata.mcp_url` as a diagnostic endpoint: `/v1/readyz` must report `dispatcher=false`, and `tools/call` must return a structured `transport-error`; do not route gateway traffic through it until `dispatch_status=ready`. Maya `commandport://` sidecars return a structured `sidecar-dispatcher-unavailable` backend envelope when `dcc_mcp_maya` is present but its sidecar dispatcher is missing, so treat that as a partial adapter install rather than a gateway routing failure. If a ready backend later returns terminal `host-died`, gateway removes that instance from the capability index plus its FileRegistry or HTTP registration row immediately; do not design adapters around waiting for stale heartbeat cleanup after a terminal host failure.
+8. When a DCC startup hook must spawn the per-DCC sidecar without importing native core, use `dcc_mcp_core.install_lifecycle.build_sidecar_command(...)` or `launch_sidecar(...)`; they produce the canonical `dcc-mcp-server sidecar` argv, keep stdio detached by default, stamp the shared registry/gateway env, and let the child ensure the daemon without blocking the DCC main process unless `no_ensure_gateway=True` or `legacy_gateway_election=True`. Do not treat a `per-dcc-sidecar` registry row alone as proof that tools are callable: the adapter must expose a supported host RPC bridge to its DCC dispatcher/skills, and startup code or installers should use `sidecar_readiness_status(...)` / `wait_for_sidecar_ready(...)` for a bounded import-light verdict before claiming the plugin is ready. Pass `probe_tool="<dcc>_diagnostics__ping"` (or CLI `--probe-tool`) when the adapter needs proof that one real `tools/call` reaches the host dispatcher; probe failures keep the verdict unready until timeout. Operators should see `dispatch_ready=True` from those helpers, `query_runtime_state(...)`, `GET /admin/api/workers`, or raw `metadata.dispatch_status=ready` plus `metadata.mcp_url`; failed host RPC startup stays registered with `dispatch_ready=False`, `metadata.dispatch_status=unavailable`, and `failure_stage` / `failure_reason` for diagnostics. A failed sidecar may still publish `metadata.mcp_url` as a diagnostic endpoint: `/v1/readyz` must report `dispatcher=false`, and `tools/call` must return a structured `transport-error`; do not route gateway traffic through it until `dispatch_status=ready`. `stub://` is test-only and stays `dispatch_status=unavailable` by default even when it connects; never use it as adapter startup proof. Maya `commandport://` sidecars return a structured `sidecar-dispatcher-unavailable` backend envelope when `dcc_mcp_maya` is present but its sidecar dispatcher is missing, so treat that as a partial adapter install rather than a gateway routing failure. If a ready backend later returns terminal `host-died`, gateway removes that instance from the capability index plus its FileRegistry or HTTP registration row immediately; do not design adapters around waiting for stale heartbeat cleanup after a terminal host failure.
 9. If the adapter cannot share the gateway `FileRegistry`, register remotely through `POST /v1/instances/register`, refresh with `/heartbeat`, and deregister on shutdown; the gateway will expose the row as `source: "http"` in `gateway://instances` / `GET /v1/instances`, preserve `instance_short` and `mcp_url`, and route it through the same `live_instances` contract.
 10. For same-LAN convenience discovery, build with `mdns` and pair adapter-side `--advertise-mdns` with gateway-side `--discover-mdns`; treat this as a multicast discovery hint only, keep auth/TLS policy explicit, and prefer HTTP registration or relay for routed/subnet-crossing production deployments.
 11. For NAT or routed-subnet deployments, run the tunnel agent with stable `instance_id`, `capabilities_fingerprint`, `adapter_version`, and `scene` metadata, then configure the standalone gateway with `--relay-source ADMIN_URL=PUBLIC_BASE_URL`; the gateway will expose active tunnels as `source: "relay"` rows with relay details in `source_meta` after probing `/v1/healthz` through `<PUBLIC_BASE_URL>/tunnel/<tunnel_id>/mcp`.
